+++ hoomd/pytest/dummy.py
@@ -1,4 +1,5 @@
-from hoomd.meta import _Operation
+from hoomd.triggers import Trigger
+from hoomd.meta import _Operation, _TriggeredOperation
 from hoomd.parameterdicts import TypeParameterDict, RequiredArg
 from hoomd.typeparam import TypeParameter
 
@@ -7,6 +8,12 @@ class DummySimulation:
     def __init__(self):
         self.state = DummyState()
         self.operations = DummyOperations()
+        self._cpp_sys = DummySystem()
+
+
+class DummySystem:
+    def __init__(self):
+        self.dummy_list = []
 
 
 class DummyState:
@@ -50,17 +57,17 @@ class DummyCppObj:
 
 
 class DummyOperation(_Operation):
-    def __init__(self):
-        super().__init__(self)
-        self._param_dict = dict(param1=1, param2=2)
-        tp1 = TypeParameter(name='type_param', type_kind='particle_types',
-                            param_dict=TypeParameterDict(foo=1,
-                                                         bar=RequiredArg,
-                                                         baz='hello')
-                            )
-        self._add_typeparam(tp1)
+    '''Requires that user manually add param_dict and typeparam_dict items.
 
-    def attach(self, sim):
-        self._cpp_obj = DummyCppObj()
-        self._apply_param_dict()
-        self._apply_typeparam_dict(self._cpp_obj, sim)
+    This is for testing purposes.
+    '''
+    pass
+
+
+class DummyTriggeredOp(_TriggeredOperation):
+    _cpp_list_name = 'dummy_list'
+
+
+class DummyTrigger(Trigger):
+    def __call__(self, ts):
+        return True